refactor: drop dist files, lint, use async method
